logic for admin and all checks

+++ Material.BLL/Services/MaterialService.cs
@@ -26,13 +26,19 @@ public class MaterialService(IMaterialRepository materialRepository, IMapper map
         return materialModel;
     }
 
-    public async Task CreateMaterialAsync(MaterialEntityModel material, CancellationToken cancellationToken = default)
+    public async Task CreateMaterialAsync(int userId, MaterialEntityModel material,  CancellationToken cancellationToken = default)
     {
-        var materialDb = await _materialRepository.GetAll()
-            .FirstOrDefaultAsync(i => i.Name == material.Name,
-                cancellationToken);
+        var userDb = await _userRepository.GetByIdAsync(userId, cancellationToken);
+        if (userDb is null)
+            throw new UserNotFoundException($"User with this Id {userId} not found");
+        
+        if (userDb.Login is not "admin")
+            throw new NoRightException("You don't have permissions");
+        
+        var isMaterialExists = await _materialRepository.GetAll()
+            .AnyAsync(i => i.Name == material.Name, cancellationToken);
 
-        if (materialDb is not null && materialDb.Name == material.Name)
+        if (isMaterialExists)
             throw new AlreadyMaterialNameException("Material name is already used by another user");
 
         var materialDbModel = _mapper.Map<MaterialEntity>(material);
@@ -40,13 +46,19 @@ public class MaterialService(IMaterialRepository materialRepository, IMapper map
         material.Name = materialDbModel.Name;
     }
 
-
-    public async Task DeleteMaterialAsync(int id, CancellationToken cancellationToken = default)
+    public async Task DeleteMaterialAsync(int userId, int materialId,  CancellationToken cancellationToken = default)
     {
-        var materialDb = await _materialRepository.GetByIdAsync(id, cancellationToken);
+        var materialDb = await _materialRepository.GetByIdAsync(materialId, cancellationToken);
         if (materialDb is null)
-            throw new MaterialNotFoundException($"Material with this Id {id} not found");
-
+            throw new MaterialNotFoundException($"Material with this Id {materialId} not found");
+        
+        var userDb = await _userRepository.GetByIdAsync(userId, cancellationToken);
+        if (userDb is null)
+            throw new UserNotFoundException($"User with this Id {userId} not found");
+        
+        if (userDb.Login is not "admin")
+            throw new NoRightException("You don't have permissions");
+        
         await _materialRepository.DeleteMaterialAsync(materialDb, cancellationToken);
     }
 
@@ -59,7 +71,13 @@ public class MaterialService(IMaterialRepository materialRepository, IMapper map
         var userDb = await _userRepository.GetByIdAsync(userId, cancellationToken);
         if (userDb is null)
             throw new UserNotFoundException($"User with this Id {userId} not found");
-        //check for isExist ?
+        
+        var isAlreadyInFavorites = await _favoriteListMaterialRepository.GetAll()
+            .AnyAsync(f => f.UserId == userId && f.MaterialId == materialId, cancellationToken);
+
+        if (isAlreadyInFavorites)
+            throw new MaterialAlreadyInFavoritesException($"Material with Id {materialId} is already in the favorites list for user with Id {userId}");
+
         await _favoriteListMaterialRepository.AddMaterial(materialDb, userDb, cancellationToken);
     }
     
@@ -72,22 +90,29 @@ public class MaterialService(IMaterialRepository materialRepository, IMapper map
         var userDb = await _userRepository.GetByIdAsync(userId, cancellationToken);
         if (userDb is null)
             throw new UserNotFoundException($"User with this Id {userId} not found");
-        //check for isExist ?
+        
+        var isInFavorites = await _favoriteListMaterialRepository.GetAll()
+            .AnyAsync(f => f.UserId == userId && f.MaterialId == materialId, cancellationToken);
+
+        if (!isInFavorites)
+            throw new MaterialNotInFavoritesException($"Material with Id {materialId} is not in the favorites list for user with Id {userId}");
+        
         await _favoriteListMaterialRepository.RemoveMaterial(materialDb.Id, userDb.Id, cancellationToken);
     }
 
-    public async Task<MaterialEntityModel> UpdateMaterialAsync(int id, MaterialEntityModel material, CancellationToken cancellationToken = default)
+
+    public async Task<MaterialEntityModel> UpdateMaterialAsync(int materialId, MaterialEntityModel materialEntityModel, CancellationToken cancellationToken = default)
     {
-        var materialDb = await _materialRepository.GetByIdAsync(id, cancellationToken);
+        var materialDb = await _materialRepository.GetByIdAsync(materialId, cancellationToken);
         if (materialDb is null)
-            throw new MaterialNotFoundException($"Material with this Id {id} not found");
+            throw new MaterialNotFoundException($"Material with this Id {materialId} not found");
         
         foreach (var propertyMap in ReflectionHelper.WidgetUtil<MaterialEntityModel, MaterialEntity>.PropertyMap)
         {
             var userProperty = propertyMap.Item1;
             var userDbProperty = propertyMap.Item2;
         
-            var userSourceValue = userProperty.GetValue(material);
+            var userSourceValue = userProperty.GetValue(materialEntityModel);
             var userTargetValue = userDbProperty.GetValue(materialDb);
         
             if (userSourceValue != null && !ReferenceEquals(userSourceValue, "") && !userSourceValue.Equals(userTargetValue))
